feat(dropzone): replace all story with file list
